[otbn,util] Move CodeSection to shared/ directory.
The CodeSection class from check_loop.py is also useful for analyzing
control flow in other contexts; move it to shared/ so it can be imported
by other checks.
